modify avatar upoloader 127.0.0.1	www.dream-demo.com

+++ file-gateway/src/main/java/com/sparrow/file/servlet/FileUpload.java
@@ -363,7 +363,7 @@ public class FileUpload extends HttpServlet {
         attachUploadParam.setPathKey(pathKey);
         attachUploadParam.setEditor(editor);
         attachUploadParam.setContentType(status.getContentType());
-        attachUploadParam.setContentLength(status.getContentLength());
+        attachUploadParam.setContentLength(status.getContentLength().intValue());
         return attachUploadParam;
     }
 }
